blktrace: Fix replay_align 32 bit truncation
When using replay_align the aligned offset was truncated when the
original offset was greater than 2^32. Fix this by forcing the
calculation to happen with 64 bits.

+++ blktrace.c
@@ -221,7 +221,7 @@ static void ipo_bytes_align(struct thread_options *o, struct io_piece *ipo)
 	if (!o->replay_align)
 		return;
 
-	ipo->offset &= ~(o->replay_align - 1);
+	ipo->offset &= ~(o->replay_align - (uint64_t)1);
 }
